+++ slskit/__main__.py
@@ -17,7 +17,7 @@ import slskit.pillar
 import slskit.state
 import slskit.template
 from slskit import PACKAGE_NAME, VERSION
-from slskit.opts import DEFAULT_CONFIG_PATHS, DEFAULT_SNAPSHOT_PATH, Config
+from slskit.opts import DEFAULT_CONFIG_PATHS, Config
 from slskit.types import AnyDict, MinionDict
 
 
@@ -107,71 +107,10 @@ def refresh(ctx: click.Context) -> None:
         salt.runners.saltutil.sync_all()
 
 
-@cli.group(help="create and check highstate snapshots")
-@click.option(
-    "-p",
-    "--path",
-    "snapshot_path",
-    default=DEFAULT_SNAPSHOT_PATH,
-    type=Path,
-    help=f"path to snapshot file (default: {DEFAULT_SNAPSHOT_PATH})",
-)
-@click.pass_context
-def snapshot(ctx: click.Context, snapshot_path: Path) -> None:
-    ctx.obj["snapshot_path"] = snapshot_path
-
-
-@snapshot.command(help="create highstate snapshot")
-@click.pass_context
-def create(ctx: click.Context) -> None:
-    dump = _dump_highstate(ctx.obj["config"])
-    if not dump:
-        sys.exit("Failed to render snapshot")
-
-    ctx.obj["snapshot_path"].write_text(dump)
-    print(f"Snapshot saved as `{ctx.obj['snapshot_path']}`")
-
-
-@snapshot.command(help="check highstate snapshot")
-@click.pass_context
-def check(ctx: click.Context) -> None:
-    if not ctx.obj["snapshot_path"].exists():
-        sys.exit(f"Snapshot file `{ctx.obj['snapshot_path']}` not found")
-    snapshot = ctx.obj["snapshot_path"].read_text()
-
-    dump = _dump_highstate(ctx.obj["config"])
-    if not dump:
-        sys.exit("Failed to render snapshot")
-
-    if snapshot != dump:
-        _display_diff(snapshot, dump)
-        sys.exit(
-            "There are some changes not present in the snapshot. "
-            "Run `slskit snapshot create` to update the snapshot."
-        )
-
-
 def _output(minion_dict: MinionDict, config: Config) -> None:
     salt.output.display_output(minion_dict.output, opts=config.opts)
     if not minion_dict.all_valid:
         sys.exit(1)
-
-
-def _dump_highstate(config: Config) -> Optional[str]:
-    minion_ids = config.roster.keys()
-    minion_dict = slskit.state.show_highstate(minion_ids, config)
-    if not minion_dict.all_valid:
-        return None
-
-    dump = salt.utils.yaml.safe_dump(minion_dict.output, default_flow_style=False)
-    return cast(str, dump)
-
-
-def _display_diff(a: str, b: str) -> None:
-    diff = difflib.unified_diff(
-        a.splitlines(keepends=True), b.splitlines(keepends=True)
-    )
-    sys.stdout.writelines(diff)
 
 
 cli()